requirements.txt
asgiref==3.4.1
Django==3.2.8
djangorestframework==3.12.4
pytz==2021.3
sqlparse==0.4.2
psycopg2-binary==2.8.6
djangorestframework-simplejwt==4.3.0
djoser==2.1.0
requests==2.22.0
PyJWT==1.7.1